+++ tests/test_ipyvizzu.py
@@ -7,7 +7,6 @@ from normalizer import Normalizer
 from ipyvizzu import (
     PlainAnimation,
     Data,
-    Filter,
     Config,
     Style,
     Animate,
@@ -36,20 +35,6 @@ class TestAnimation(unittest.TestCase):
     def test_style_can_be_none(self):
         animation = Style(None)
         self.assertEqual({"style": None}, animation.build())
-
-    def test_filter(self):
-        animation = Filter("filter_expression")
-        self.assertEqual(
-            '{"data": {"filter": record => { return (filter_expression) }}}',
-            animation.dump(),
-        )
-
-    def test_filter_can_be_none(self):
-        animation = Filter(None)
-        self.assertEqual(
-            '{"data": {"filter": null}}',
-            animation.dump(),
-        )
 
 
 class TestMethod(unittest.TestCase):
@@ -87,49 +72,6 @@ class TestMerger(unittest.TestCase):
             self.merger.build(),
         )
 
-    def test_merge_data_and_filter(self):
-        data = Data()
-        records = [
-            ["Soul", "Hard", 91],
-            ["Soul", "Smooth", 57],
-            ["Soul", "Experimental", 115],
-        ]
-        data.add_records(records)
-        self.merger.merge(data)
-        self.merger.merge(Filter(None))
-        self.assertEqual(
-            """{"data": {"records": [["Soul", "Hard", 91], ["Soul", "Smooth", 57], ["Soul", "Experimental", 115]], "filter": null}}""",  # pylint: disable=line-too-long
-            self.merger.dump(),
-        )
-
-    def test_only_one_data_can_be_merged(self):
-        data1 = Data()
-        records = [
-            ["Soul", "Hard", 91],
-            ["Soul", "Smooth", 57],
-            ["Soul", "Experimental", 115],
-        ]
-        data1.add_records(records)
-        self.merger.merge(data1)
-        self.merger.merge(Filter(None))
-
-        data2 = Data()
-        record = ["Rock", "Hard", 96]
-        data2.add_record(record)
-        self.assertRaises(ValueError, self.merger.merge, data2)
-
-    def test_only_one_filter_can_be_merged(self):
-        data1 = Data()
-        records = [
-            ["Soul", "Hard", 91],
-            ["Soul", "Smooth", 57],
-            ["Soul", "Experimental", 115],
-        ]
-        data1.add_records(records)
-        self.merger.merge(data1)
-        self.merger.merge(Filter(None))
-        self.assertRaises(ValueError, self.merger.merge, Filter(None))
-
     def test_only_different_type_of_animation_can_be_merged(self):
         self.merger.merge(Config({"channels": {"label": {"attach": ["Popularity"]}}}))
         self.assertRaises(
@@ -153,6 +95,20 @@ class TestMerger(unittest.TestCase):
 class TestData(unittest.TestCase):
     def setUp(self):
         self.data = Data()
+
+    def test_filter(self):
+        data = Data.filter("filter_expression")
+        self.assertEqual(
+            '{"data": {"filter": record => { return (filter_expression) }}}',
+            data.dump(),
+        )
+
+    def test_filter_can_be_none(self):
+        data = Data.filter(None)
+        self.assertEqual(
+            '{"data": {"filter": null}}',
+            data.dump(),
+        )
 
     def test_record(self):
         self.data.add_record(["Pop", "Hard", 114])