+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 ### Added
 
 - Added CONTRIBUTING.md.
-- Added `Filter` animation.
+- Added `Data.filter()` and `Data().set_filter()` methods.
 - Added `Data.add_records()` function to add records' list in one step.
 
 ## [0.5.0] - 2022-03-12